+++ tests/testthat/test-count_points_in_polygons.R
@@ -1,6 +1,7 @@
 data_sf <- suppressWarnings(
   sf::st_centroid(sf::read_sf(system.file("shape/nc.shp", package = "sf")))
 )
+data_sf$wt <- runif(nrow(data_sf), max = 1000)
 data_df <- as.data.frame(sf::st_drop_geometry(data_sf))
 
 polygons <- create_grid(data = data_sf, cell_size = 0.1)
@@ -34,6 +35,19 @@ test_that("error if `polygons` is not an SF object containing polygons", {
   ))
 })
 
+test_that("error if `weights` is not NULL or the name of a colum in `points`", {
+  expect_error(count_points_in_polygons(
+    points = data_sf,
+    polygons = polygons,
+    wt = character()
+  ))
+  expect_error(count_points_in_polygons(
+    points = data_sf,
+    polygons = polygons,
+    wt = "blah"
+  ))
+})
+
 
 
 # CHECK OUTPUTS ----------------------------------------------------------------
@@ -48,6 +62,14 @@ test_that("function produces an SF tibble", {
 
 test_that("output object has the required column names", {
   expect_equal(names(result), c("n", "geometry"))
+  expect_equal(
+    names(count_points_in_polygons(
+      points = data_sf,
+      polygons = polygons,
+      weights = "wt"
+    )),
+    c("n", "sum", "geometry")
+  )
 })
 
 test_that("columns in output have the required types", {